requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt[crypto]>=5.4,<6
PyJWT[crypto]>=2.9,<3
python-dotenv>=1.0,<2
gunicorn>=23.0,<24
